Change the notion of async "test" to "software."
The async simulator is a software backend for the async code that just calls
into wolfcrypt to handle crypto ops instead of some truly async hardware device.
This is useful for testing, but it also provides a nice way for non-blocking
crypto algorithms in wolfCrypt to be used by wolfSSL's TLS/DTLS layer, as the
non-blocking coding patterns map neatly onto the async patterns. With
non-blocking TLS/DTLS being a "real-world" use case of the async simulator,
this commit changes all the language in the async code from "test" to
"software."